requirements.txt
# Web framework
fastapi
uvicorn[standard]
jinja2

# LLM / LangChain
langchain-ollama
langchain-core

# Validación de datos
pydantic

# Generación de documentos
python-docx
markdown2
weasyprint
openpyxl

# Base de datos
tinydb

# Autenticación
bcrypt
PyJWT
